+++ crates/icu/src/macros/start.rs
@@ -86,6 +86,7 @@ macro_rules! icu_start_root {
 
         fn __icu_shared_setup() {
             ::icu::__icu_load_config!();
+            ::icu::memory::CanisterPool::start();
             ::icu::memory::CycleTracker::start();
             ::icu::state::wasm::WasmRegistry::import(WASMS);
             icu_setup();